move proeprty selection to own component and some fixes
